--- ui/src/main/java/io/snabble/sdk/ui/payment/CreditCardInputActivity.kt
@@ -1,17 +1,0 @@
-package io.snabble.sdk.ui.payment
-
-import androidx.fragment.app.Fragment
-import io.snabble.sdk.ui.BaseFragmentActivity
-
-class CreditCardInputActivity : BaseFragmentActivity() {
-    companion object {
-        const val ARG_PROJECT_ID = CreditCardInputView.ARG_PROJECT_ID
-        const val ARG_PAYMENT_TYPE = CreditCardInputView.ARG_PAYMENT_TYPE
-    }
-
-    override fun onCreateFragment(): Fragment {
-        val fragment = CreditCardInputFragment()
-        fragment.arguments = intent.extras
-        return fragment
-    }
-}
