+++ ui/src/main/java/io/snabble/sdk/ui/payment/fiserv/data/CountryItemsRepositoryImpl.kt
@@ -1,0 +1,12 @@
+package io.snabble.sdk.ui.payment.fiserv.data
+
+import io.snabble.sdk.ui.payment.fiserv.data.country.LocalCountryItemsDataSource
+import io.snabble.sdk.ui.payment.fiserv.domain.CountryItemsRepository
+import io.snabble.sdk.ui.payment.fiserv.domain.model.country.CountryItem
+
+internal class CountryItemsRepositoryImpl(
+    private val localCountryItemsDataSource: LocalCountryItemsDataSource,
+) : CountryItemsRepository {
+
+    override fun loadCountryItems(): List<CountryItem> = localCountryItemsDataSource.loadCountries()
+}
