+++ ui/src/main/java/io/snabble/sdk/ui/payment/fiserv/data/country/LocalCountryItemsDataSource.kt
@@ -1,0 +1,8 @@
+package io.snabble.sdk.ui.payment.fiserv.data.country
+
+import io.snabble.sdk.ui.payment.fiserv.domain.model.country.CountryItem
+
+internal interface LocalCountryItemsDataSource {
+
+    fun loadCountries(): List<CountryItem>
+}
