+++ ui/src/main/java/io/snabble/sdk/ui/payment/fiserv/domain/CountryItemsRepository.kt
@@ -1,0 +1,8 @@
+package io.snabble.sdk.ui.payment.fiserv.domain
+
+import io.snabble.sdk.ui.payment.fiserv.domain.model.country.CountryItem
+
+internal interface CountryItemsRepository {
+
+    fun loadCountryItems(): List<CountryItem>
+}
