+++ ui/src/main/java/io/snabble/sdk/ui/payment/fiserv/domain/FiservRepository.kt
@@ -1,0 +1,9 @@
+package io.snabble.sdk.ui.payment.fiserv.domain
+
+import io.snabble.sdk.PaymentMethod
+import io.snabble.sdk.ui.payment.fiserv.data.FiservCardRegisterUrls
+
+internal interface FiservRepository {
+
+    suspend fun sendUserData(customerInfo: CustomerInfo, paymentMethod: PaymentMethod): Result<FiservCardRegisterUrls>
+}
