+++ ui/src/main/java/io/snabble/sdk/ui/payment/fiserv/domain/model/country/CountryItem.kt
@@ -1,0 +1,3 @@
+package io.snabble.sdk.ui.payment.fiserv.domain.model.country
+
+internal data class CountryItem(val displayName: String, val code: String, val stateItems: List<StateItem>?)
