+++ ui/src/main/java/io/snabble/sdk/ui/payment/fiserv/widget/TextInput.kt
@@ -1,0 +1,78 @@
+package io.snabble.sdk.ui.payment.fiserv.widget
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.text.KeyboardActions
+import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.OutlinedTextField
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextFieldDefaults
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.TextRange
+import androidx.compose.ui.text.input.ImeAction
+import androidx.compose.ui.text.input.KeyboardCapitalization
+import androidx.compose.ui.text.input.TextFieldValue
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+
+@Composable
+internal fun TextInput(
+    modifier: Modifier = Modifier,
+    value: String,
+    label: String,
+    isError: Boolean = false,
+    onValueChanged: (String) -> Unit,
+    keyboardOptions: KeyboardOptions = KeyboardOptions(
+        imeAction = ImeAction.Next,
+        capitalization = KeyboardCapitalization.Words
+    ),
+    keyboardActions: KeyboardActions,
+) {
+    var textFieldValue by remember {
+        mutableStateOf(TextFieldValue(text = value, selection = TextRange(value.length, value.length)))
+    }
+    Column(modifier = modifier, verticalArrangement = Arrangement.spacedBy(4.dp)) {
+        OutlinedTextField(
+            modifier = modifier,
+            value = textFieldValue,
+            onValueChange = {
+                onValueChanged(it.text)
+                textFieldValue = it
+            },
+            textStyle = MaterialTheme.typography.bodyLarge,
+            label = {
+                Text(
+                    text = label,
+                    fontSize = 17.sp
+                )
+            },
+            keyboardOptions = keyboardOptions,
+            keyboardActions = keyboardActions,
+            isError = isError,
+            maxLines = 1,
+            singleLine = true,
+            colors = TextFieldDefaults.defaultColors(),
+        )
+    }
+}
+
+@Composable
+private fun TextFieldDefaults.defaultColors() = colors(
+    focusedContainerColor = MaterialTheme.colorScheme.background,
+    unfocusedContainerColor = MaterialTheme.colorScheme.background,
+    focusedIndicatorColor = MaterialTheme.colorScheme.primary,
+    unfocusedIndicatorColor = MaterialTheme.colorScheme.onSurface,
+    focusedLabelColor = MaterialTheme.colorScheme.primary,
+    errorContainerColor = MaterialTheme.colorScheme.background,
+    unfocusedLabelColor = MaterialTheme.colorScheme.onSurface,
+    errorCursorColor = Color.Red,
+    errorIndicatorColor = Color.Red,
+    errorLabelColor = Color.Red
+)
