--- ui/src/main/java/io/snabble/sdk/ui/payment/telecash/data/CountryItemsRepositoryImpl.kt
@@ -1,12 +1,0 @@
-package io.snabble.sdk.ui.payment.telecash.data
-
-import io.snabble.sdk.ui.payment.telecash.data.country.LocalCountryItemsDataSource
-import io.snabble.sdk.ui.payment.telecash.domain.CountryItemsRepository
-import io.snabble.sdk.ui.payment.telecash.domain.model.country.CountryItem
-
-internal class CountryItemsRepositoryImpl(
-    private val localCountryItemsDataSource: LocalCountryItemsDataSource,
-) : CountryItemsRepository {
-
-    override fun loadCountryItems(): List<CountryItem> = localCountryItemsDataSource.loadCountries()
-}
