--- ui/src/main/java/io/snabble/sdk/ui/payment/telecash/data/country/LocalCountryItemsDataSource.kt
@@ -1,8 +1,0 @@
-package io.snabble.sdk.ui.payment.telecash.data.country
-
-import io.snabble.sdk.ui.payment.telecash.domain.model.country.CountryItem
-
-internal interface LocalCountryItemsDataSource {
-
-    fun loadCountries(): List<CountryItem>
-}
