--- ui/src/main/java/io/snabble/sdk/ui/payment/telecash/domain/CountryItemsRepository.kt
@@ -1,8 +1,0 @@
-package io.snabble.sdk.ui.payment.telecash.domain
-
-import io.snabble.sdk.ui.payment.telecash.domain.model.country.CountryItem
-
-internal interface CountryItemsRepository {
-
-    fun loadCountryItems(): List<CountryItem>
-}
