--- ui/src/main/java/io/snabble/sdk/ui/payment/telecash/domain/model/country/CountryItem.kt
@@ -1,3 +1,0 @@
-package io.snabble.sdk.ui.payment.telecash.domain.model.country
-
-internal data class CountryItem(val displayName: String, val code: String, val stateItems: List<StateItem>?)
