+++ ui/src/main/res/layout/snabble_fragment_cardinput_fiserv.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<io.snabble.sdk.ui.payment.fiserv.FiservInputView xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@+id/user_payment_method_view"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent" />
